requirements.txt
# Core dependencies
python-telegram-bot==20.3
python-dotenv>=1.0.0

# Data processing (optional - only if needed for future features)
# numpy==2.2.6
# pandas==2.3.3

# Image processing (optional - only if needed for QR code generation)
# pillow==12.0.0
# qrcode==8.2

# HTTP client dependencies (required by python-telegram-bot)
anyio==4.11.0
certifi==2025.10.5
exceptiongroup==1.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
sniffio==1.3.1
typing_extensions==4.15.0

# Date/time utilities (required by python-telegram-bot)
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0
